requirements.txt
fastapi
uvicorn[standard]
requests
PyGithub
sentence-transformers
faiss-cpu
transformers
jinja2
python-dotenv
openai
apache-airflow-client
pydantic-settings
pydantic>=2.0
langchain
langchain-community
langchain-openai
langchainhub
neo4j
python-multipart
langchain-openai
